Add colours to website and aligns page
Website looked plain and boring so i added a warm tan colour to the
background and a darker brown to h2 for dividing title and image from
contents. also added colour to the home button to make it easier to find.

Also made it so website alignes more center on wider monitors.